requirements.txt
mindsdb_sdk 
pandas 
numpy
crewai
matplotlib
requests
python-dotenv
rasterio